cmg/go: Set GCC/Clang -arch flag during CGO build
On macOS BigSur 11.1 running on Apple Silicon the default C compiler
supports building for multiple architectures: arm64 and amd64. (This
is true of clang on other platforms as well.)

When running Go 1.16 in in native ARM mode, and building a package
into an x86 executable using GOARCH=amd64, then compilation fails on
gcc_darwin_amd64.c and it embedded asm directives due to the clang
compiler defaulting to arm64 mode.

To resolve this, add a check if the compiler supports the "-arch" [1]
argument. If so, set it to GOARCH's equivalent GCC/LLVM value.

[1] https://clang.llvm.org/docs/CommandGuide/clang.html#target-selection-options
    Many Clang cross-compilation howtos recommend using "-target", but
    this requires also mapping the OS and ABI names to the LLVM/GCC
    equivalent. Most versions of GCC do not support the "-arch" flag,
    however Apple's GCC adds support. Search for "Darwin Options" at
    https://www.man7.org/linux/man-pages/man1/gcc.1.html